refactor(apollo): extract context definition to own file
