+++ index.js
@@ -12,9 +12,7 @@ const { Action } = require('./models/Action')
 const { ApolloServer } = require('apollo-server')
 const { GraphQLJSONObject } = require('graphql-type-json')
 const { Model } = require('objection')
-const { queryResolvers, mutationResolvers } = require('./resolvers')
-const cookie = require('cookie')
-const jwt = require('jsonwebtoken')
+const { queryResolvers, mutationResolvers, setContext } = require('./resolvers')
 const Knex = require('knex')
 const knexConfig = require('./knexfile.js')
 const typeDefs = require('./schema')
@@ -57,28 +55,7 @@ const server = new ApolloServer({
   resolvers,
   introspection: process.env.INTROSPECTION,
   playground: process.env.PLAYGROUND,
-  context: ({ req }) => {
-    let decoded
-
-    try {
-      const authCookieName = process.env.SSO_COOKIE_NAME
-      const cookies = cookie.parse(req.headers.cookie)
-      const authToken = cookies[authCookieName]
-      decoded = jwt.verify(authToken, process.env.SSO_JWT_SECRET)
-    } catch (err) {
-      // eslint-disable-next-line
-      console.error(err)
-
-      return { User: {} }
-    }
-
-    // TODO: create the user entry within our service database
-    // const user = User.findOrCreateFromJWT(decoded);
-    // User.findByExternalId(decoded.external_id)
-    // const user = User.create({external_id, ...rest })
-
-    return { User: decoded }
-  }
+  context: setContext
 })
 
 server.listen().then(({ url }) => {
