+++ resolvers/index.js
@@ -2,6 +2,7 @@ const { AuthenticationError } = require('apollo-server')
 const { Campaign } = require('../models/Campaign')
 const { User } = require('../models/User')
 const { UserAction } = require('../models/UserAction')
+const setContext = require('./context')
 
 const queryResolvers = {
   /**
@@ -78,6 +79,7 @@ const mutationResolvers = {
 }
 
 module.exports = {
+  setContext,
   queryResolvers,
   mutationResolvers
 }
